C++ : - Snake Game
Snake is the common name for a video game concept where the player maneuvers a line which grows in length, with the line itself being a primary obstacle. The player controls a dot, square, or object on a bordered plane. As it moves forward, it leaves a trail behind, resembling a moving snake. In some games, the end of the trail is in a fixed position, so the snake continually gets longer as it moves.

Features :
Program created  with functions ( avoided OOP and data structures ) 
Game terminates only if user press 'q' or snake touches its own tail.
Displayed score which updates after snake eats fruit. 

Next step 
I Will develop the snake game using direct x framework.